+++ packages/cli-tools/bin/common.ts
@@ -1,4 +1,19 @@
-function envOptions(program) {
+import * as commander from 'commander'
+import { StreamrClientOptions } from 'streamr-client'
+
+export interface EnvironmentOptions {
+    dev?: boolean
+    stg?: boolean
+    wsUrl?: string
+    httpUrl?: string 
+}
+
+export interface AuthenticationOptions {
+    privateKey?: string
+    apiKey?: string
+}
+
+export function envOptions(program: commander.Command) {
     return program
         .option('--dev', 'use pre-defined development environment')
         .option('--stg', 'use pre-defined staging environment')
@@ -6,20 +21,20 @@ function envOptions(program) {
         .option('--http-url <url>', 'alternative http url to use')
 }
 
-function authOptions(program) {
+export function authOptions(program: commander.Command) {
     return program
         .option('--private-key <key>', 'use an Ethereum private key to authenticate')
         .option('--api-key <key>', 'use an API key to authenticate (deprecated)')
 }
 
-function exitWithHelpIfArgsNotBetween(program, min, max) {
+export function exitWithHelpIfArgsNotBetween(program: commander.Command, min: number, max: number) {
     if (program.args.length < min || program.args.length > max) {
         program.help()
     }
 }
 
-function formStreamrOptionsWithEnv({ dev, stg, wsUrl, httpUrl, privateKey, apiKey }) {
-    const options = {}
+export function formStreamrOptionsWithEnv({ dev, stg, wsUrl, httpUrl, privateKey, apiKey }: EnvironmentOptions & AuthenticationOptions) {
+    const options: StreamrClientOptions = {}
 
     if (dev && stg) {
         console.error('flags --dev and --stg cannot be enabled at the same time')
@@ -29,6 +44,11 @@ function formStreamrOptionsWithEnv({ dev, stg, wsUrl, httpUrl, privateKey, apiKe
     if (dev) {
         options.url = 'ws://localhost/api/v1/ws'
         options.restUrl = 'http://localhost/api/v1'
+        options.storageNode = {
+            // "broker-node-storage-1" on Docker environment
+            address: '0xde1112f631486CfC759A50196853011528bC5FA0',
+            url: 'http://10.200.10.1:8891'
+        }
     } else if (stg) {
         options.url = 'wss://staging.streamr.com/api/v1/ws'
         options.restUrl = 'https://staging.streamr.com/api/v1/'
@@ -59,8 +79,8 @@ function formStreamrOptionsWithEnv({ dev, stg, wsUrl, httpUrl, privateKey, apiKe
     return options
 }
 
-function createFnParseInt(name) {
-    return (str) => {
+export function createFnParseInt(name: string) {
+    return (str: string) => {
         const n = parseInt(str, 10)
         if (isNaN(n)) {
             console.error(`${name} must be an integer (was "${str}")`)
@@ -68,12 +88,4 @@ function createFnParseInt(name) {
         }
         return n
     }
-}
-
-module.exports = {
-    envOptions,
-    authOptions,
-    exitWithHelpIfArgsNotBetween,
-    formStreamrOptionsWithEnv,
-    createFnParseInt
 }