+++ packages/cli-tools/bin/streamr-stream.ts
@@ -1,15 +1,17 @@
 #!/usr/bin/env node
-const program = require('commander')
+import { program } from 'commander'
+import pkg from '../package.json'
 
 program
-    .version(require('../package.json').version)
+    .version(pkg.version)
     .usage('<command> [<args>]')
-    .description('command line tools for interacting with Streamr https://streamr.network')
-    .command('generate', 'generate JSON data')
+    .description('stream subcommands')
     .command('subscribe', 'subscribe to a stream')
     .command('publish', 'publish to a stream')
     .command('list', 'fetch a list of streams')
     .command('show', 'info about a stream')
     .command("create", "create a new stream")
     .command('resend', 'request resend of a stream')
+    .command('grant-permission', 'grant permission')
+    .command('revoke-permission', 'revoke permission')
     .parse(process.argv)