+++ packages/cli-tools/src/create.ts
@@ -1,10 +1,11 @@
-const StreamrClient = require('streamr-client')
+import { StreamProperties, StreamrClient } from 'streamr-client'
 
-module.exports = function create(body, streamrOptions) {
+export const create = (body: any, streamrOptions: StreamProperties) => {
     const options = { ...streamrOptions }
 
     const client = new StreamrClient(options)
     client.createStream(body).then((stream) => {
+        // @ts-expect-error
         console.info(JSON.stringify(stream.toObject(), null, 2))
         process.exit(0)
     }).catch((err) => {